Signed/unsigned warnings cleanup
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17948 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/gssapi/krb5/k5unseal.c
@@ -80,8 +80,8 @@ kg_unseal_v1(context, minor_status, ctx, ptr, bodysize, message_buffer,
     char *data_ptr;
     krb5_timestamp now;
     unsigned char *plain;
-    int cksum_len = 0;
-    int plainlen;
+    unsigned int cksum_len = 0;
+    size_t plainlen;
     int direction;
     krb5_ui_4 seqnum;
     OM_uint32 retval;
